requirements.txt
streamlit
pdfminer.six
nltk
tabula-py
